requirements.txt
# LangChain
langchain
langchain-core
langchain-community
langchain-experimental

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
sentence-transformers

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
streamlit

# pydantic
pydantic
pydantic[email]

# to visualize chains
grandalf

# pyPDFLoader
pyPDF

# WebBaseLoader
bs4

# Chroma vector database
langchain-chroma
chromadb

# retrievers
faiss-cpu
tiktoken
wikipedia

youtube-transcript-api

# tools
ddgs # duckduckgo-search